Move query_classifier.py into the classifier module, update tests and tutorials
