+++ haystack/classifier/__init__.py
@@ -1,1 +1,3 @@
-from haystack.classifier.farm import FARMClassifier
+from haystack.classifier.document_classifier.farm import FARMClassifier
+from haystack.classifier.query_classifier.sklearn import SklearnQueryClassifier
+from haystack.classifier.query_classifier.transformers import TransformersQueryClassifier
