+++ haystack/classifier/document_classifier/base.py
